requirements.txt
python-telegram-bot==13.15
Flask==2.2.5
gunicorn==21.2.0